+++ StrassenAlgorithm/StrassenAlgorithm.h
@@ -11,9 +11,9 @@ class StrassenAlgorithm {
         static vector<vector<int>> splitAndSolve(vector<vector<int>>, vector<vector<int>>, int);
         static vector<vector<int>> mergeMatrix(vector<vector<int>>, vector<vector<int>>, int, int);
     public:
-        static double strassen(const string&, const string&);
-        static double divideAndConquer(const string&, const string&);
-        static double bruteForce(const string&, const string&);
+        static int strassen(const string&, const string&);
+        static int divideAndConquer(const string&, const string&);
+        static int bruteForce(const string&, const string&);
 };
 
 #endif //ALGORITHMS_STRASSENALGORITHM_H
